% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_correct_annual.R
\name{bias_correct_annual}
\alias{bias_correct_annual}
\title{Annual correction of monthly bias-corrected data}
\usage{
bias_correct_annual(
  simulated,
  bias_corrected,
  start_date_simulated,
  end_date_simulated
)
}
\arguments{
\item{simulated}{Simulated time series}

\item{bias_corrected}{Monthly bias-corrected time series}

\item{start_date_simulated}{Start date of simulated time series. For example: "1929-01-01"}

\item{end_date_simulated}{End date of simulated time series. For example: "1929-01-01"}
}
\description{
This function makes sure that the magnitude of annual simulated streamflow is consistent with the simulations
}
