% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_correct_daily.R
\name{bias_correct_daily}
\alias{bias_correct_daily}
\title{Bias correction of daily streamflow}
\usage{
bias_correct_daily(
  observed,
  simulated,
  start_date_observed,
  end_date_observed,
  start_date_simulated,
  end_date_simulated,
  time_step,
  date_type,
  disaggregation_method
)
}
\arguments{
\item{observed}{Observed time series}

\item{simulated}{Simulated time series}

\item{start_date_observed}{Start date of observed time series. For example: "1929-01-01"}

\item{end_date_observed}{End date of observed time series. For example: "1929-01-01"}

\item{start_date_simulated}{Start date of simulated time series. For example: "1929-01-01"}

\item{end_date_simulated}{Start date of simulated time series. For example: "1929-01-01"}

\item{time_step}{Time step of the observed and simulated time-series. Can be "day" or "month"}

\item{date_type}{Water year (WY) or Julian Year (JY)}

\item{disaggregation_method}{Can be "knn" or "scaling_coeff"}
}
\description{
This function first applies monthly and annual bias correction and then disaggregates to daily
}
