% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_correct_monthly.R
\name{bias_correct_monthly}
\alias{bias_correct_monthly}
\title{Bias correction of monthly streamflow}
\usage{
bias_correct_monthly(
  observed,
  simulated,
  start_date_observed,
  end_date_observed,
  start_date_simulated,
  end_date_simulated,
  time_step,
  date_type
)
}
\arguments{
\item{observed}{Observed time series.}

\item{simulated}{Simulated time series.}

\item{start_date_observed}{Start date of observed time series. For example: "1929-01-01".}

\item{end_date_observed}{End date of observed time series. For example: "1929-01-01".}

\item{start_date_simulated}{Start date of simulated time series. For example: "1929-01-01".}

\item{end_date_simulated}{End date of simulated time series. For example: "1929-01-01".}

\item{time_step}{Time step of the observed and simulated time-series. Can be "day" or "month".}

\item{date_type}{Water year (WY) or Julian Year (JY).}
}
\description{
This function uses the quantile mapping technique to calculate monthly bias-corrected data
}
