% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ccr_monthly}
\alias{ccr_monthly}
\title{ccr_monthly}
\usage{
ccr_monthly(fcst, obs, fcst.out = fcst, fc.time, fcout.time = fc.time, ...)
}
\arguments{
\item{fcst}{array of forecast values (nyear, nlead, nens)}

\item{obs}{array of observations (nyear, nlead)}

\item{fcst.out}{array of forecast values to which bias correction
should be applied (defaults to \code{fcst})}

\item{fc.time}{forecast times as R-dates for monthly aggregation}

\item{fcout.time}{forecast time for array to which bias correction is applied
for back compatibility with leave-one-out cross-validation in \code{\link{debias}}}

\item{...}{additional arguments for compatibility with other bias correction methods}
}
\description{
Computes climate conserving recalibration on smoothed monthly means
}
\keyword{util}

