% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ccrlm.R
\name{ccrlm}
\alias{ccrlm}
\title{Climate conserving recalibration}
\usage{
ccrlm(fcst, obs, fcst.out = fcst, ...)
}
\arguments{
\item{fcst}{n x m x k array of n lead times, m forecasts, of k ensemble
members}

\item{obs}{n x m matrix of veryfing observations}

\item{fcst.out}{array of forecast values to which bias correction should be
applied (defaults to \code{fcst})}

\item{...}{additional arguments for compatibility with other bias correction
  methods}
}
\description{
Climate conserving recalibration with correction factors for each lead time
without smoothing (application to daily series not recommended).
}
\examples{
nveri <- 100
ncal <- 10
nn <- nveri + ncal
nens <- 5
signal <- rnorm(nn, sd=5)
sdfrac <- 2
fcst <- array(rnorm(nn*nens, sd=sdfrac), c(1, nn, nens)) + 0.5 * signal
obs <- array(rnorm(nn, sd=1), c(1, nn)) + signal
fcst.debias <- biascorrection:::ccrlm(fcst[,1:ncal,,drop=F], obs[,1:ncal,drop=F],
                                      fcst.out=fcst[, ncal + 1:nveri,,drop=F])
f.rmse <- sqrt(apply((obs[,ncal + 1:nveri, drop=F] - apply(fcst.debias, 2, mean))**2, 1, mean))
f.rmse2 <- sqrt(apply((obs[,ncal + 1:nveri,drop=F] - apply(fcst[,ncal + 1:nveri,,drop=F], 2, mean))**2, 1, mean))
f.sd <- sqrt(mean(apply(fcst.debias, 2, sd)**2))
f.sd2 <- sqrt(mean(apply(fcst[,ncal + 1:nveri,,drop=F], 2, sd)**2))
f.sd / f.rmse ## should be exactly 1
f.sd2 / f.rmse2 ## can be anything
mean(fcst.debias - obs[1,]) ## should be 0 (rounding errors)
}
\keyword{util}

