% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conditional.R
\name{comb}
\alias{comb}
\alias{combRecal}
\title{Bias comb on forecast}
\usage{
comb(...)

combRecal(...)
}
\arguments{
\item{...}{additional arguments for compatibility with other bias correction
  methods}

\item{fcst}{n x m x k array of n lead times, m forecasts, of k ensemble
members}

\item{obs}{n x m matrix of veryfing observations}

\item{fcst.out}{array of forecast values to which bias correction should be
applied (defaults to \code{fcst})}

\item{span}{the parameter which controls the degree of smoothing (see
\code{\link{loess}})}
}
\description{
Compute calibration for biases that are comb on ensemble mean forecast
}
\examples{
## initialise forcast observation pairs
seasonal <- sin(seq(0,4,length=215))
signal <- outer(seasonal, rnorm(30), '+')
fcst <- array(rnorm(215*30*51), c(215, 30, 15)) +
  2*c(signal)
obs <- array(rnorm(215*30, mean=2), c(215, 30)) +
  signal
fc.time <- outer(1:215, 1981:2010, function(x,y) as.Date(paste0(y, '-11-01')) - 1 + x)
fcst.debias <- biascorrection:::comb(fcst[,1:20,],
obs[,1:20], fcst.out=fcst, fc.time=fc.time[],1:20], fcout.time=fc.time, span=0.5)
}
\seealso{
linmod
}
\keyword{util}

