% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cor}
\alias{cor}
\title{Unbiased correlation}
\usage{
cor(x, y, method = "pearson", ...)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}

\item{method}{a character string indicating which correlation coefficient
is to be computed. The unbiased version is only available for
\code{"pearson"} (the default).}

\item{...}{additional arguments passed to \code{\link{cor}}

The correlation function introduced here uses the bias correction from
Olkin and Pratt (1958) as shown below where \eqn{\hat{\rho}} is the
unbiased correlation estimator and \eqn{r} is the standard sample
correlation. The unbiased correlation is
\deqn{\hat{\rho} = r \left[ 1 + \frac{1 - r^2}{2(n - 3)} \right]}}
}
\description{
This function computes the unbiased population estimate of the correlation
}
\examples{
x <- rnorm(20)
y <- rnorm(20) + 0.5*x
biascorrection:::cor(x,y)
stats::cor(x,y)
}
\references{
Olkin, I. and Pratt, J.W. (1958). Unbiased estimation
    of certain correlation coefficients. Annals of Mathematical
    Statistics, 29, 201-211.
}

