% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_quantiles.R
\name{create_quantiles}
\alias{create_quantiles}
\title{This function is called to create monthly quantiles}
\usage{
create_quantiles(df_input, n_year)
}
\arguments{
\item{n_year}{}
}
\description{
This function is called from the "bias_correct_monthly" functions and converst a data frame of input data into its monthly quantiles.
}
