% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiasApply.R
\name{debiasApply}
\alias{debiasApply}
\title{Apply Bias Correction to Large Ensemble Forecast Data Sets}
\usage{
debiasApply(fcst, obs, method = "unbias", fcst.out = fcst, ...,
  atomic = FALSE)
}
\arguments{
\item{fcst}{array of ensemble forecast data with arbitrary dimensions. The 
final dimensions contain lead times, forecast instances (usually years) and
ensemble members in this order.}

\item{obs}{array of observation used for calibration. Dimensions correspond
to the \code{fcst} array without the last dimension.}

\item{method}{character string with bias correction method name}

\item{fcst.out}{array of ensemble forecast data that is to be calibrated. All
but the forecast instance and ensemble dimensions have to be the same as
for \code{fcst}.}

\item{...}{Additional parameters passed on to \code{\link{debias}}}

\item{atomic}{logical, should debiasing be applied to each forecast instance
separately (in contrast to jointly over lead times)?}
}
\description{
This wrapper allows to apply the bias correction methods 
  supplied with \code{\link{debias}} to large arrays of ensemble forecast and
  observation data.
}
\keyword{util}

