% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initial.R
\name{initial}
\alias{initial}
\title{Bias dependent on initial condition}
\usage{
initial(fcst, obs, fcst.out = fcst, span = min(1, 31/nrow(fcst)), ...)
}
\arguments{
\item{fcst}{n x m x k array of n lead times, m forecasts, of k ensemble members}

\item{obs}{n x m matrix of veryfing observations}

\item{fcst.out}{array of forecast values to which bias correction
should be applied (defaults to \code{fcst})}

\item{span}{the parameter which controls the degree of smoothing (see \code{\link{loess}})}

\item{...}{additional arguments for compatibility with other bias correction methods}
}
\description{
Mean De-biasing in Dependence of Initial Conditions
}
\details{
This bias correction method assumes that the time-dependent mean bias depends on the
initial conditions. The method loosely follows the ideas outlined in Fuckar et al. (2014),
but in contrast to their approach, we use the forecast and observed conditions at the
first day of the forecast as a proxy for the initial condition. Thereby, individual ensemble
members have varying bias correction depending on their respective initial conditions.
}
\examples{
## initialise forcast observation pairs
fcst <- array(rnorm(215*30*51), c(215, 30, 51)) +
0.5*sin(seq(0,4,length=215)) +
rep(seq(0,1,length=30), each=215)
obs <- array(rnorm(215*30, mean=2), c(215, 30)) +
sin(seq(0,4, length=215)) +
rep(seq(0,3,length=30), each=215)
fc.time <- outer(1:215, 1981:2010, function(x,y) as.Date(paste0(y, '-11-01')) - 1 + x)
fcst.debias <- biascorrection:::initial(fcst[,1:20,],
obs[,1:20], fcst.out=fcst, fc.time=fc.time, span=0.5)
}
\keyword{util}

