% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_methods.R
\name{list_methods}
\alias{list_methods}
\title{List calibration methods}
\usage{
list_methods()
}
\description{
List calibration methods available in \code{biascorrection}
}
\examples{
list_methods()
}

