% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{monthly}
\alias{monthly}
\title{Daily Calibration with Monthly Mean}
\usage{
monthly(fcst, obs, fcst.out = fcst, fc.time, fcout.time = fc.time, ...)
}
\arguments{
\item{fcst}{n x m x k array of n lead times, m forecasts, of k ensemble members}

\item{obs}{n x m matrix of veryfing observations}

\item{fcst.out}{array of forecast values to which bias correction
should be applied (defaults to \code{fcst})}

\item{fc.time}{forecast times as R-dates for monthly aggregation}

\item{fcout.time}{forecast time for array to which bias correction is applied
for back compatibility with leave-one-out cross-validation in \code{\link{debias}}}

\item{...}{additional arguments for compatibility with other bias correction methods}
}
\description{
Computes monthly mean de-biasing
}
\keyword{util}

