% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mul}
\alias{mul}
\title{Multiplicative de-biasing}
\usage{
mul(fcst, obs, fcst.out = fcst, ...)
}
\arguments{
\item{fcst}{n x m x k array of n lead times, m forecasts, of k ensemble
members}

\item{obs}{n x m matrix of veryfing observations}

\item{fcst.out}{array of forecast values to which bias correction should be
applied (defaults to \code{fcst})}

\item{...}{additional arguments for compatibility with other bias correction
  methods}
}
\description{
Computes multiplicative de-biasing without smoothing
}
\details{
The bias corrected forecast is scaled by the lead-time dependent
  ratio of observed to forecast climatology, where the observed and the
  forecast climatologies are smoothed using a loess smoothing.
}
\examples{
## initialise forcast observation pairs
signal <- outer(1.5 + sin(seq(0,4,length=215)), rnorm(30)**2, '*')
fcst <- array(rnorm(length(signal)*15)**2, c(dim(signal), 15)) * c(signal)
obs <- rnorm(length(signal), mean=1.4)**2 * signal
fcst.debias <- biascorrection:::mul(fcst[,1:20,], obs[,1:20], fcst.out=fcst)
}
\seealso{
smoothobs_mul smooth_mul unbias
}
\keyword{util}

