% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qqmap_semi}
\alias{qqmap_semi}
\title{qqmap_semi}
\usage{
qqmap_semi(fcst, obs, fcst.out = fcst, span = min(1, 31/nrow(fcst)), ...)
}
\arguments{
\item{fcst}{array of forecast values (nyear, nlead, nens)}

\item{obs}{array of observations (nyear, nlead)}

\item{fcst.out}{array of forecast values to which bias correction
should be applied (defaults to \code{fcst})}

\item{span}{the parameter which controls the degree of qqmaping (see \code{\link{loess}})}

\item{...}{additional arguments for compatibility with other bias correction methods}
}
\description{
Computes bias correction with semi-parametric quantile mapping
}
\details{
The quantile-quantile mapping implemented here is based on the assumption that
the observation data are approximately normally distributed. In contrast, quantiles
for the forecasts are directly estimated based on the assumption that there are enough
degrees of freedom to estimate the quantiles directly. This may strongly affect forecasts
in regions with skill.
}
\examples{
## initialise forcast observation pairs
fcst <- array(runif(215*30*51), c(215, 30, 51)) +
0.5*sin(seq(0,4,length=215))
obs <- array(rnorm(215*30, mean=2), c(215, 30)) +
sin(seq(0,4, length=215))
fcst.debias <- biascorrection:::qqmap_semi(fcst[,1:20,], obs[,1:20], fcst.out=fcst, span=0.5)
}
\keyword{util}

