% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sloess.R
\name{sloess}
\alias{sloess}
\title{Compute loess fit}
\usage{
sloess(x, ...)
}
\arguments{
\item{x}{input vector}

\item{...}{additional parameters handed over to loess}
}
\description{
This function computes a loess fit on the given vector if more than 5 non-missing
values are available. Else, the raw values are returned.
}
\keyword{util}

