% Generated by roxygen2 (4.0.2): do not edit by hand
\name{smooth}
\alias{smooth}
\title{Smooth mean de-biasing}
\usage{
smooth(fcst, obs, fcst.out = fcst, span = min(1, 31/nrow(fcst)), ...)
}
\arguments{
\item{fcst}{n x m x k array of n lead times, m forecasts, of k ensemble members}

\item{obs}{n x m matrix of veryfing observations}

\item{fcst.out}{array of forecast values to which bias correction
should be applied (defaults to \code{fcst})}

\item{span}{the parameter which controls the degree of smoothing (see \code{\link{loess}})}

\item{...}{additional arguments for compatibility with other bias correction methods}
}
\description{
Computes mean de-biasing with loess smoothing
}
\keyword{util}

