% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smoothobs_mul.R
\name{smoothobs_mul}
\alias{smoothobs_mul}
\title{Multiplicative de-biasing with smoothed obs. climatology}
\usage{
smoothobs_mul(fcst, obs, fcst.out = fcst, span = min(1, 31/nrow(fcst)), ...)
}
\arguments{
\item{fcst}{n x m x k array of n lead times, m forecasts, of k ensemble
members}

\item{obs}{n x m matrix of veryfing observations}

\item{fcst.out}{array of forecast values to which bias correction should be
applied (defaults to \code{fcst})}

\item{span}{the parameter which controls the degree of smoothing (see
\code{\link{loess}})}

\item{...}{additional arguments for compatibility with other bias correction
  methods}
}
\description{
Computes multiplicative de-biasing with loess smoothing (obs. only)
}
\details{
The bias corrected forecast is scaled by the lead-time dependent
  ratio of observed to forecast climatology, where the observed climatology
  is smoothed using a loess smoothing.
}
\examples{
## initialise forcast observation pairs
signal <- outer(1.5 + sin(seq(0,4,length=215)), rnorm(30)**2, '*')
fcst <- array(rnorm(length(signal)*15)**2, c(dim(signal), 15)) * c(signal)
obs <- rnorm(length(signal), mean=1.4)**2 * signal
fcst.debias <- biascorrection:::smoothobs_mul(fcst[,1:20,], obs[,1:20], fcst.out=fcst, span=0.5)
}
\seealso{
smooth_mul smoothobs
}
\keyword{util}

