% Generated by roxygen2 (4.0.2): do not edit by hand
\name{smoothobs_scale}
\alias{smoothobs_scale}
\title{smoothobs_scale}
\usage{
smoothobs_scale(fcst, obs, fcst.out = fcst, span = min(1, 31/nrow(fcst)),
  ...)
}
\arguments{
\item{fcst}{n x m x k array of n lead times, m forecasts, of k ensemble members}

\item{obs}{n x m matrix of veryfing observations}

\item{fcst.out}{array of forecast values to which bias correction
should be applied (defaults to \code{fcst})}

\item{span}{the parameter which controls the degree of smoothing (see \code{\link{loess}})}

\item{...}{additional arguments for compatibility with other bias correction methods}
}
\description{
Computes mean de-biasing with loess smoothing (obs. only) and adjusts variance
}
\examples{
## initialise forcast observation pairs
fcst <- array(rnorm(215*30*51, mean=3, sd=0.2), c(215, 30, 51)) +
0.5*sin(seq(0,4,length=215))
obs <- array(rnorm(215*30, mean=2), c(215, 30)) +
sin(seq(0,4, length=215))
fcst.debias <- biascorrection:::smoothobs_scale(fcst[,1:20,], obs[,1:20], fcst.out=fcst, span=0.5)
}
\keyword{util}

