% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unbias}
\alias{unbias}
\title{unbias}
\usage{
unbias(fcst, obs, fcst.out = fcst, ...)
}
\arguments{
\item{fcst}{n x m x k array of n lead times, m forecasts, of k ensemble members}

\item{obs}{n x m matrix of veryfing observations}

\item{fcst.out}{array of forecast values to which bias correction
should be applied (defaults to \code{fcst})}

\item{...}{additional arguments for compatibility with other bias correction methods}
}
\description{
Computes simple mean de-biasing
}
\keyword{util}

