% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ccr}
\alias{ccr}
\title{Climate conserving recalibration}
\usage{
ccr(fcst, obs, fcst.out = fcst, ...)
}
\arguments{
\item{fcst}{n x m x k array of n lead times, m forecasts, of k ensemble members}

\item{obs}{n x m matrix of veryfing observations}

\item{fcst.out}{array of forecast values to which bias correction
should be applied (defaults to \code{fcst})}

\item{...}{additional arguments for compatibility with other bias correction methods}
}
\description{
Climate conserving recalibration with correction factors for each lead
time without smoothing (application to daily series not recommended).
}
\examples{
fcst <- array(rnorm(3000*1*51, mean=1, sd=rep(seq(0.5,2, length=3000), each=1)),
c(1, 3000, 51)) + 0.5*sin(seq(0,4,length=1))
obs <- array(rnorm(3000, mean=2), c(1, 3000)) + sin(seq(0,4, length=1))
fcst.debias <- biascorrection:::ccr(fcst, obs)
f.rmse <- sqrt(apply((obs - apply(fcst.debias, 2, mean))**2, 1, mean))
f.sd <- sqrt(mean(apply(fcst.debias, 2, sd)**2))
f.sd / f.rmse ## should be exactly 1
mean(fcst.debias - obs[1,]) ## should be 0 (rounding errors)
}
\keyword{util}

