% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccr.R
\name{ccr}
\alias{ccr}
\title{Climate Conserving Recalibration}
\usage{
ccr(fcst, obs, fcst.out = fcst, type = c("calibration", "prediction"), ...)
}
\arguments{
\item{fcst}{n x m x k array of n lead times, m forecasts, of k ensemble 
members}

\item{obs}{n x m matrix of veryfing observations}

\item{fcst.out}{array of forecast values to which bias correction should be 
applied (defaults to \code{fcst})}

\item{type}{if set to "prediction", additional inflation of the spread as in 
linear regression (see details)}

\item{...}{additional arguments for compatibility with other bias correction 
methods}
}
\description{
Climate conserving recalibration with correction factors for each lead time 
without smoothing (application to daily series not recommended).
}
\details{
In \code{calibration} mode, the climate conserving recalibration 
  (CCR) follows Weigel \emph{et al.} (2008). In \code{prediction} mode, the 
  CCR spread correction is expanded to take into account additional 
  uncertainty from the signal calibration following linear regression theory.
  I.e. the inflation factor \code{s.pred} is
  
  \deqn{s.pred = s \sqrt{1 + 1/n + f_0^2 / \sum{f_j^2}}}{s.pred = s * sqrt(1
  + 1/n + fo^2 / sum(fj^2))}
  
  where \eqn{f_0}{fo} is the ensemble mean forecast anomaly that is to be 
  adjusted, and \eqn{\sum{f_j^2}}{sum(fj^2)} is the sum of the squared 
  ensemble mean forecast anomalies in the calibration set. \eqn{n} is the 
  number of forecast instances in the calibration set.
}
\examples{
fcst <- array(rnorm(3000*1*51, mean=1, sd=rep(seq(0.5,2, length=3000), each=1)), 
c(1, 3000, 51)) + 0.5*sin(seq(0,4,length=1))
obs <- array(rnorm(3000, mean=2), c(1, 3000)) + sin(seq(0,4, length=1))
fcst.debias <- biascorrection:::ccr(fcst, obs)
f.rmse <- sqrt(apply((obs - apply(fcst.debias, 2, mean))**2, 1, mean))
f.sd <- sqrt(mean(apply(fcst.debias, 2, sd)**2))
f.sd / f.rmse ## should be exactly 1
mean(fcst.debias - obs[1,]) ## should be 0 (rounding errors)

}
\references{
Weigel, A., M. Liniger and C. Appenzeller (2008). Seasonal 
  Ensemble Forecasts: Are Recalibrated Single Models Better than Multimodels?
  \emph{Monthly Weather Review}, 137(4), 1460-1479.
}
\keyword{util}
