% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comb.R
\name{comb}
\alias{comb}
\alias{combRecal}
\title{Conditional Bias With Linear Time Trend}
\usage{
comb(...)

combRecal(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{linmod}}}
}
\description{
Compute calibration for biases that are conditional on ensemble mean forecast
and exhibit a linear time trend
}
\examples{
## initialise forcast observation pairs
seasonal <- sin(seq(0,4,length=215))
signal <- outer(seasonal, rnorm(30), '+')
fcst <- array(rnorm(215*30*51), c(215, 30, 15)) +
  2*c(signal)
obs <- array(rnorm(215*30, mean=2), c(215, 30)) +
  signal
fc.time <- outer(1:215, 1981:2010, function(x,y) as.Date(paste0(y, '-11-01')) - 1 + x)
fcst.debias <- biascorrection:::comb(fcst[,1:20,],
obs[,1:20], fcst.out=fcst, fc.time=fc.time[,1:20], fcout.time=fc.time, span=0.5)
}
\seealso{
linmod
}
\keyword{util}

