% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_methods.R
\name{list_methods}
\alias{list_methods}
\alias{print.biascorrection}
\title{List Calibration Methods}
\usage{
list_methods()

\method{print}{biascorrection}(x, ...)
}
\arguments{
\item{x}{data frame of methods to be used with \code{\link{debias}}}

\item{...}{additional parameters for consistency with generic method}
}
\description{
\code{list_methods} lists calibration methods available in \code{biascorrection}
and \code{print.biascorrection} is used to nicely format these for display.
}
\examples{
list_methods()

}
