% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/month.R
\name{month}
\alias{month}
\title{Compute the monthly mean}
\usage{
month(x, time)
}
\arguments{
\item{x}{input array (2-d or 3-d, i.e. obs or forecasting)}

\item{time}{dates of input array (2-d, e.g. lead time x years)}
}
\description{
Compute the monthly mean
}
\keyword{util}

