% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth}
\alias{smooth}
\alias{smoothRecal}
\alias{smoothobs}
\alias{smoothobsRecal}
\alias{unbias}
\alias{unbiasRecal}
\title{Mean De-biasing With Smoothing of Daily Climatology}
\usage{
unbias(...)

smoothobs(...)

smooth(...)

unbiasRecal(...)

smoothobsRecal(...)

smoothRecal(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{linmod}}}
}
\description{
Computes mean de-biasing (with and without loess smoothing)
}
\keyword{util}

