% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trend}
\alias{trend}
\title{trend}
\usage{
trend(fcst, obs, fcst.out = fcst, fc.time, fcout.time = fc.time,
  span = min(1, 31/nrow(fcst)), ...)
}
\arguments{
\item{fcst}{array of forecast values (nyear, nlead, nens)}

\item{obs}{array of observations (nyear, nlead)}

\item{fcst.out}{array of forecast values to which bias correction
should be applied (defaults to \code{fcst})}

\item{fc.time}{forecast times as R-dates for trend aggregation}

\item{fcout.time}{forecast time for array to which bias correction is applied
for back compatibility with leave-one-out cross-validation in \code{\link{debias}}}

\item{span}{the parameter which controls the degree of smoothing (see \code{\link{loess}})}

\item{...}{additional arguments for compatibility with other bias correction methods}
}
\description{
Computes mean de-biasing with linear time trend
}
\details{
This bias correction method assumes that the bias can be decomposed into a stationary
seasonal cycle (as in method \code{\link{smoothobs}}) and a linear time trend estimated
from the residuals.
}
\examples{
## initialise forcast observation pairs
fcst <- array(rnorm(215*30*51), c(215, 30, 51)) +
0.5*sin(seq(0,4,length=215)) +
rep(seq(0,1,length=30), each=215)
obs <- array(rnorm(215*30, mean=2), c(215, 30)) +
sin(seq(0,4, length=215)) +
rep(seq(0,3,length=30), each=215)
fc.time <- outer(1:215, 1981:2010, function(x,y) as.Date(paste0(y, '-11-01')) - 1 + x)
fcst.debias <- trend(fcst[,1:20,], obs[,1:20], fcst.out=fcst, fc.time=fc.time, span=0.5)
}
\keyword{util}

