% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trend.R
\name{trend}
\alias{trend}
\alias{trendRecal}
\title{Bias with linear time trend}
\usage{
trend(...)

trendRecal(...)
}
\arguments{
\item{...}{additional arguments for compatibility with other bias correction methods}

\item{fcst}{n x m x k array of n lead times, m forecasts, of k ensemble members}

\item{obs}{n x m matrix of veryfing observations}

\item{fcst.out}{array of forecast values to which bias correction
should be applied (defaults to \code{fcst})}

\item{fc.time}{forecast times as R-dates for trend aggregation}

\item{fcout.time}{forecast time for array to which bias correction is applied
for back compatibility with leave-one-out cross-validation in \code{\link{debias}}}

\item{span}{the parameter which controls the degree of smoothing (see \code{\link{loess}})}
}
\description{
Computes mean de-biasing with linear time trend
}
\details{
This bias correction method assumes that the bias can be decomposed
into a stationary seasonal cycle (as in method \code{\link{smoothobs}})
and a linear time trend estimated from the residuals.
}
\examples{
## initialise forcast observation pairs
fcst <- array(rnorm(215*30*51), c(215, 30, 51)) +
0.5*sin(seq(0,4,length=215)) +
rep(seq(0,1,length=30), each=215)
obs <- array(rnorm(215*30, mean=2), c(215, 30)) +
sin(seq(0,4, length=215)) +
rep(seq(0,3,length=30), each=215)
fc.time <- outer(1:215, 1981:2010, function(x,y) as.Date(paste0(y, '-11-01')) - 1 + x)
fcst.debias <- biascorrection:::trend(fcst[,1:20,],
obs[,1:20], fcst.out=fcst, fc.time=fc.time[],1:20], fcout.time=fc.time, span=0.5)
}
\seealso{
linmod
}
\keyword{util}

